refactor: :recycle: move socket connection to config
