+++ backend/src/index.ts
@@ -1,21 +1,15 @@
 import express, { Request, Response, Application } from "express"
 import dotenv from "dotenv"
 import exampleRouter from "./routes/example.routes"
-import { createServer } from "http"
-import { Server } from "socket.io"
 import cors from "cors"
+import { connectToSocket } from "./configs/socketConnection"
 
 // For env File
 dotenv.config()
 
 const app: Application = express()
 const port = process.env.PORT ?? 8000
-const server = createServer(app)
-const io = new Server(server, {
-  cors: {
-    origin: "*"
-  }
-})
+const server = connectToSocket(app)
 
 app.use(
   cors({
@@ -25,19 +19,6 @@ app.use(
 
 app.get("/HelloWorld", (req: Request, res: Response) => {
   res.send("Hello World")
-})
-
-io.on("connection", (socket) => {
-  console.log("New user connected")
-
-  socket.on("sendMessage", (message) => {
-    console.log(message)
-    socket.broadcast.emit("receiveMessage", message)
-  })
-
-  socket.on("disconnect", () => {
-    console.log("User disconnected")
-  })
 })
 
 server.listen(port, () => {
